Move extra types from MainProcess to their own files
