+++ PotatoChipMine/EmptyCommand.cs
@@ -1,0 +1,8 @@
+using PotatoChipMine.Models;
+
+namespace PotatoChipMine
+{
+    public class EmptyCommand : UserCommand
+    {
+    }
+}
