+++ PotatoChipMine/EventLog.cs
@@ -1,0 +1,9 @@
+namespace PotatoChipMine
+{
+    public class EventLog
+    {
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public string Processed { get; set; }
+    }
+}
