+++ PotatoChipMine/GameMode.cs
@@ -1,0 +1,10 @@
+namespace PotatoChipMine
+{
+    public enum GameMode
+    {
+        Lobby = 1,
+        Store = 2,
+        Mining = 3,
+        ControlRoom = 4
+    }
+}
